fix warnings
git-svn-id: http://localhost:9111/private/safir_open/safir_sdk_core/trunk@591 9dd7cf02-b8c9-4ff8-a4a5-e7fda3bdd973

+++ src/dose/dose_test.ss/Sequencer/ActionSender.h
@@ -102,7 +102,7 @@ public:
         }
         else
         {
-            SendInternal(binary,msg->Partner().GetVal().GetRawValue());
+            SendInternal(binary,static_cast<int>(msg->Partner().GetVal().GetRawValue()));
         }
 
         SleepyTime(msg->ActionKind());
